Robustness loss added with gaussian perturbations without optimization

+++ dice_ml_x/utils/helpers.py
@@ -15,7 +15,7 @@ from sklearn.model_selection import train_test_split
 from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import FunctionTransformer, OneHotEncoder
 
-import dice_ml
+import dice_ml_x
 
 
 def load_adult_income_dataset(only_train=True):
@@ -27,7 +27,8 @@ def load_adult_income_dataset(only_train=True):
     # Download the adult dataset from https://archive.ics.uci.edu/static/public/2/adult.zip as a zip folder
     outdirname = 'adult'
     zipfilename = outdirname + '.zip'
-    urlretrieve('https://archive.ics.uci.edu/static/public/2/adult.zip', zipfilename)
+    if not os.path.isfile(zipfilename):
+        urlretrieve('https://archive.ics.uci.edu/static/public/2/adult.zip', zipfilename)
     with zipfile.ZipFile(zipfilename, 'r') as unzip:
         unzip.extractall(outdirname)
 
@@ -185,42 +186,42 @@ def load_custom_testing_dataset_regression():
 
 
 def get_adult_income_modelpath(backend='TF1'):
-    pkg_path = dice_ml.__path__[0]
+    pkg_path = dice_ml_x.__path__[0]
     model_ext = '.h5' if 'TF' in backend else ('.pth' if backend == 'PYT' else '.pkl')
     modelpath = os.path.join(pkg_path, 'utils', 'sample_trained_models', 'adult'+model_ext)
     return modelpath
 
 
 def get_custom_dataset_modelpath_pipeline():
-    pkg_path = dice_ml.__path__[0]
+    pkg_path = dice_ml_x.__path__[0]
     model_ext = '.sav'
     modelpath = os.path.join(pkg_path, 'utils', 'sample_trained_models', 'custom'+model_ext)
     return modelpath
 
 
 def get_custom_vars_dataset_modelpath_pipeline():
-    pkg_path = dice_ml.__path__[0]
+    pkg_path = dice_ml_x.__path__[0]
     model_ext = '.sav'
     modelpath = os.path.join(pkg_path, 'utils', 'sample_trained_models', 'custom_vars'+model_ext)
     return modelpath
 
 
 def get_custom_dataset_modelpath_pipeline_binary():
-    pkg_path = dice_ml.__path__[0]
+    pkg_path = dice_ml_x.__path__[0]
     model_ext = '.sav'
     modelpath = os.path.join(pkg_path, 'utils', 'sample_trained_models', 'custom_binary'+model_ext)
     return modelpath
 
 
 def get_custom_dataset_modelpath_pipeline_multiclass():
-    pkg_path = dice_ml.__path__[0]
+    pkg_path = dice_ml_x.__path__[0]
     model_ext = '.sav'
     modelpath = os.path.join(pkg_path, 'utils', 'sample_trained_models', 'custom_multiclass'+model_ext)
     return modelpath
 
 
 def get_custom_dataset_modelpath_pipeline_regression():
-    pkg_path = dice_ml.__path__[0]
+    pkg_path = dice_ml_x.__path__[0]
     model_ext = '.sav'
     modelpath = os.path.join(pkg_path, 'utils', 'sample_trained_models', 'custom_regression'+model_ext)
     return modelpath
@@ -246,7 +247,7 @@ def get_base_gen_cf_initialization(data_interface, encoded_size, cont_minx, cont
     #                      Check dependencies.
     from torch import optim
 
-    from dice_ml.utils.sample_architecture.vae_model import CF_VAE
+    from dice_ml_x.utils.sample_architecture.vae_model import CF_VAE
 
     # Dataset for training Variational Encoder Decoder model for CF Generation
     df = data_interface.normalize_data(data_interface.one_hot_encoded_data)
